WIP: This branch is no longer necessary
I'm pushing this WIP work that I did quite a while back. It's messy and contains
some old files (like backup and history files which were being used for debugging
and also to keep track or prior approaches while making change).

A complex and fleshed out testing infrastructure is unnecessary work and complexity,
given that the production system is extremely redundant, with many copies of the same
canister, the deployment workflow is to deploy to one or two canisters first and observe
their behaviour in the wild for several weeks before deploying to additional canisters.
This effectively allows for testing in the wild without impacting the behaviour of the
known neurons, given the large amount of redundancy and resilience.

Given this, a local testing environment (for a system that isn't intended to be changed
except for minor adjustments if there are any breaking API changes) is unnecessary distraction
given that there's lots of other stuff that needs developing for the IC that I would prefer
to focus on. Maybe I'll revisit this in the future when I'm more freed up, as the same sort
of testing framework could be used to good effect in other projects